+++ src/parse/create.js
@@ -1,4 +1,6 @@
-import { getConstructorArgs } from '../args.js'
+import setErrorClass from 'set-error-class'
+
+import { unpackConstructorArgs } from '../args.js'
 
 // Custom error classes might throw due to missing parameters in the
 // constructor. When this happens, we silently revert to `Error`.
@@ -7,13 +9,9 @@ export const createError = function (
   classes,
 ) {
   const ErrorClass = getErrorClass(name, classes)
-  const args = getConstructorArgs(constructorArgs, message, ErrorClass)
-
-  try {
-    return new ErrorClass(...args)
-  } catch {
-    return new Error(message)
-  }
+  return Array.isArray(constructorArgs)
+    ? createErrorWithArgs(message, ErrorClass, constructorArgs)
+    : createErrorWithoutArgs(message, ErrorClass)
 }
 
 // Custom error classes can be passed to the `classes` option.
@@ -44,3 +42,43 @@ const BUILTIN_CLASSES = new Set([
   // Browser-only
   'DOMException',
 ])
+
+// With `constructorArgs`, tries to call the constructor, or default to `Error`.
+// Static properties are still set after initialization
+//  - So `constructorArgs` is only useful with custom `classes`
+//  - They might override properties set by the constructor
+//     - In case those properties have been modified after initialization
+//     - However, this means properties that had not been modified after
+//       initialization but contain non-JSON-safe values are not preserved
+//         - Unfortunately, we cannot know whether this is the case or not
+//         - Also, this is simpler to understand as: non-JSON-safe values are
+//           generally not serializable, either in options or static properties,
+//           even when set by constructor
+//            - Exception: properties where both:
+//               - Value is set in constructor
+//               - Key cannot be serialized, i.e. it is symbol, private, upper
+//                 scope or non-enumerable
+const createErrorWithArgs = function (message, ErrorClass, constructorArgs) {
+  const args = unpackConstructorArgs(constructorArgs, message)
+
+  try {
+    return new ErrorClass(...args)
+  } catch {
+    return new Error(message)
+  }
+}
+
+// Without `constructorArgs`, we default to not calling the `constructor` and
+// setting the prototype manually.
+// This means some logic performed in the constructor might be missing:
+//  - Setting properties
+//     - With non-enumerable or symbol keys
+//     - Or that are not JSON-serializable
+//  - Setting variables in an upper scope, including global variables
+// This can sometimes be worked around by setting those separately after
+// parsing, e.g. using an `init()` method.
+const createErrorWithoutArgs = function (message, ErrorClass) {
+  const error = new Error(message)
+  setErrorClass(error, ErrorClass)
+  return error
+}